fix(build): Werror strictness round 1 - sign-conversion, truncation, libssh 0.10 compat
- str_reserve: inverted overflow check broke every dynamic string (found by
  first full -Werror build; 87 unit tests now pass locally with Werror ON)
- ssh_channel_get_exit_state guarded by LIBSSH_VERSION (0.10 fallback to
  deprecated exit_status); int8_t base64 nibble widened via uint8_t
- glib guint/guint16 casts; snprintf precision for path buffers
- theme CSS printf arg count now matches placeholders exactly

+++ forwarding/fwd.c
@@ -429,7 +429,7 @@ bool ec_fwd_stop(EcFwdManager* m, int index)
         char port[16];
         snprintf(port, sizeof port, "%d", r->rule.bind_port);
         GSocketClient* c = g_socket_client_new();
-        GSocketConnection* self = g_socket_client_connect_to_host(c, "127.0.0.1", r->rule.bind_port, NULL, NULL);
+        GSocketConnection* self = g_socket_client_connect_to_host(c, "127.0.0.1", (guint16)r->rule.bind_port, NULL, NULL);
         if (self) g_object_unref(self);
         g_object_unref(c);
         g_thread_join(r->thread);
